

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.r
\name{CRMiss}
\alias{CRMiss}



\title{Abundance estimation from capture-recapture data when covariates are missing at random}



\usage{

CRMiss(cov.bin = NULL, cov.miss, cap.times, ind.obs,
K, level = 0.95, M = 100)

}



\arguments{

\item{cov.bin}{a vector, the completely observed binary covariate.}

\item{cov.miss}{a vector, the covariate which is subject to missing.}

\item{cap.times}{a vector, where each element means the number of times of being captured
for captured individuals.}

\item{ind.obs}{a vector, where elements take 0 or 1.
1 means \code{cov.miss} is not missing and 0 otherwise.}

\item{K}{the number of capture occasions.}

\item{level}{the nominal confidence level.}

\item{M}{the number of imputed datasets when using the multiple imputation method.}

}



\value{

A list containing the following components:

\item{nu.est}{a matrix whose rows represent four methods,
1st column point estimates and last two columns interval estimates of the abundance.
See \link{Details} for the meanings of MELE, IPW, MI and CC methods.}

\item{beta.est}{a matrix whose rows represent four methods,
and the columns the estimates of parameters in the Huggins-Alho's model.}

\item{alpha.el}{a list where \code{gamma0.el} means MELE of the probability of being never captured.
If \code{cov.bin} is NULL, \code{gammajk.el} is a vector with elements meaning the MELE of the probability of being captured k times for k=1,...,K.
If \code{cov.bin} is not NULL, \code{gammajk.el} is a matrix with elements meaning the MELE of the probability of being captured k and \code{cov.bin=j} for j=0,1 and k=1,...,K.}

\item{loglike.el}{the value of the maximum profile empirical log-likelihood.}

\item{converge}{a value indicating whether the algorithm converges when maximizing the profile empirical log-likelihood. 1 means that the algorithm converges and 0 otherwise.}

}



\description{
Four estimation methods for the abundance are carried out under capture-recapture models when covariates are missing at random.
}



\details{
When covariates are missing at random, four estimation methods are carried out under capture-recapture models.

First method is a new maximum empirical likelihood estimation (MELE) method.
Refer to Liu et al. (2019+).

Second method is based on the naive maximum empirical likelihood derived from the complete cases (CC).
See Liu et al. (2017).

The last two are based on the inverse probability weighting (IPW) and multiple imputation (MI) methods.
See Lee et al. (2016).
}




\references{

Lee, S.-M., W.-H. Hwang, and J. de Dieu Tapsoba (2016).
Estimation in closed capture–recapture models when covariates are missing at random.
\emph{Biometrics} \strong{72}(4), 1294--1304.

Liu, Y., P. Li, and J. Qin (2017).
Maximum empirical likelihood estimation for abundance in a closed
population from capture-recapture data.
\emph{Biometrika} \strong{104}(3), 527--543.

Liu, Y., Y. Liu, P. Li, and L. Zhu (2019+).
Maximum likelihood abundance estimation from capture-recapture data when covariates are missing at random.
Submitted.
}




\examples{
## Analysis results of the prinia data set.

## Model 1
outcomes.model1 <- CRMiss(cov.bin = prinia$fat.index, cov.miss = prinia$tail.length, cap.times = prinia$number.of.capture,
                          ind.obs = (prinia$tail.length != 0) + 0, K = 17, level = 0.95, M = 100)

## Model 2
outcomes.model2 <- CRMiss(cov.miss = prinia$tail.length, cap.times = prinia$number.of.capture,
                       ind.obs = (prinia$tail.length != 0) + 0, K = 17, level = 0.95, M = 100)

}









