

% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.r
\name{loglike.null}
\alias{loglike.null}



\title{The maximum empirical log-likelihood when the coefficient of the completely observed covariate is 0}



\usage{

loglike.null(cov.bin, cov.miss, cap.times, ind.obs, K)

}



\arguments{

\item{cov.bin}{a vector, the completely observed binary covariate.}

\item{cov.miss}{a vector, the covariate which is subject to missing.}

\item{cap.times}{a vector, where each element means the number of times of being captured 
for captured individuals.}

\item{ind.obs}{a vector, where elements take 0 or 1. 
1 means \code{cov.miss} is not missing and 0 otherwise.}

\item{K}{the number of capture occasions.}

}



\value{

The maximum empirical log-likelihood under the null hypothesis \eqn{H_0}.

}



\description{To use the empirical likelihood ratio test under the null hypothesis \eqn{H_0}: the coefficient of the completely observed covariate is 0,
it is necessary to calculate the maximum empirical log-likelihood under \eqn{H_0}.}


\references{

Liu, Y., Y. Liu, P. Li, and L. Zhu (2019+).
Maximum likelihood abundance estimation from capture-recapture data when covariates are missing at random.
Submitted.

}




\examples{
## Analysis results of the prinia data set.

## Model 1 
outcomes.model1 <- CRMiss(cov.bin = prinia$fat.index, cov.miss = prinia$tail.length, cap.times = prinia$number.of.capture, 
                          ind.obs = (prinia$tail.length != 0) + 0, K = 17, level = 0.95, M = 100)

## Test the null hypothesis \eqn{H_0: \beta_{X^*} = 0} in Model 1
like.null <- loglike.null(cov.bin = prinia$fat.index, cov.miss = prinia$tail.length, cap.times = prinia$number.of.capture, 
                           ind.obs = (prinia$tail.length != 0) + 0, K = 17)

## The empirical likelihood ratio test statistic
lrt.stat <- 2 * (outcomes.model1$loglike.el -  like.null)
lrt.stat                

## p-value
1 - pchisq(lrt.stat, 1)

}









