
\name{prinia}
\alias{prinia}


\title{Yellow-bellied prinia data set}


\arguments{
This data set is collected in Hongkong during 17 weeks from January to April in 1993. It contains 163 observations and consists of 6 columns:

\item{id}{the unique identifier.}

\item{number.of.capture}{the number of times an individual is captured.}

\item{tail.length}{the individual covariate (50mm~91.25mm) with 41 missing values (0).}

\item{fat.index}{ the fat indicator. 0 means that the bird is not fat and 1 means fat, which respectively corresponds to level 1 and level 2-4 in original data.}

\item{wing}{the wing length (43mm~49mm).}

\item{wing.index}{the wing length indicator. 0 means that the wing length ranges from 37mm to 45.5mm,
and 1 means that the wing length is greater than 45.5mm.}

}

\references{
Liu, Y., Y. Liu, P. Li, and L. Zhu (2020+). Maximum likelihood abundance estimation from capture-recapture data when covariates are missing at random. Submitted.
}




